BUGFIX:GCC warnings

+++ ftkGUI/StatisticsToolbar.cpp
@@ -203,7 +203,7 @@ QList<double> StatisticsToolbar::SortColumn(vtkAbstractArray *Column, int rows)
 		newValue = Column->GetVariantValue(j).ToDouble();
 		this->myList->push_back(newValue);
 
-		for(unsigned int i=1; (int)i< this->selectedRowNumbers.size(); i++)
+		for(unsigned int i=1; i< this->selectedRowNumbers.size(); i++)
 		{
 			j = this->selectedRowNumbers.at(i);//set row #
 			newValue = Column->GetVariantValue(j).ToDouble();
